+++ pkg/operator/operands/admission/resources.go
@@ -136,6 +136,10 @@ func (a *Admission) serviceForKAIConfig(
 	service.Spec.SessionAffinity = v1.ServiceAffinityNone
 	service.Spec.Type = v1.ServiceTypeClusterIP
 
+	if kaiConfig.Spec.Global != nil {
+		common.ApplyServiceAnnotations(service, kaiConfig.Spec.Global.ServiceAnnotations)
+	}
+
 	return []client.Object{service}, nil
 }