+++ pkg/operator/operands/queue_controller/resources.go
@@ -118,6 +118,10 @@ func (q *QueueController) serviceForKAIConfig(
 		"app": q.BaseResourceName,
 	}
 
+	if kaiConfig.Spec.Global != nil {
+		common.ApplyServiceAnnotations(service, kaiConfig.Spec.Global.ServiceAnnotations)
+	}
+
 	return []client.Object{service}, nil
 }